+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Master Summoner` OP ability.
 - Added UMM option to allow Isekai dialogue when Isekai Protagonist is disabled.
 - Added additional harem dialogue for Anevia and Irabeth.
+- Added dialogue for Finnean.
 - Added `None` option for `Hax` selection.
 - Added missing DLC spells to Isekai Protagonist spellbook.
 # Changed